[!!!][TASK] Improve templating of academic persons
The HTML templates should be as modular as
reasonably possible and should also look appealing
with only bootstrap included. Furthermore they should
be semantically optimal fot them to be as accessible
and slim as possible. Therefore the templates were
adjusted and split into reasonable partials. Furthermore
a few labels inside the locallang files got fixed as their
key was incorrect.