[TASK] Add test for missing `slug` generation with extbase model
Creating a new profile using the extbase model and persist it
using the extbase persistence manager does not take care of
creating the `slug` field value or values for the `*_alpha`
columns like done using `DataHandler` and in combination with
the provided `DataHandler` hook.

`EXT:academic_person` also provides a command to create profiles
for all frontend users that do not have a profile yet but should
have one and the provided `AbstractProfileFactory` along with
default `ProfileFactory` implementation does not take care of
the aforementioned issues.

These handlings needs a general concept as it may also occurs
for other models and are in any case a unregonized design and
implementation flaw for a long time now.

This change adds a functional test for the `Profile` to make
the above mentioned issues crystal clear along with a couple
of to-do comments.