[BUGFIX] ACE-33: Display available genders as select field
The `Profile->gender` field is defined as `TCAtype=select`
with a defined set of available options already used for
TYPO3 Backend data management, but the implementation for
`EXT:academic_persons_edit` frontend form handling used
a simple text input field diverting FE and BE handling.

This change alignes FE and BE handling with:

* Available genders are read from the TCA definition of
  the `Profile->gender` field, only respecting simply
  items and not evaluating itemProcFunc or mirroring
  custom `FormEngine` node handling and proccessing.

* Pass the available genders down and render `gender`
  field as select box now.

Note that server side validation of selected gender
against available genders are not processed yet and
will be considerd and/or implemented in a dedicated
change.